requirements.txt
python-binance==1.0.19
click==8.1.7
colorama==0.4.6
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3